added REMOTE_AUTOMATIC_VIEW (false= index.html does not get password) added custom dir support. if you add files to custom/ they will be copied to target, and if exist custom_root.sh will be executed

+++ docker-entrypoint.sh
@@ -5,18 +5,23 @@ if test "$1" = "ssh"; then
   shift
 
   if test -z "$REMOTE_VIEWONLY_PASSWORD" -o "$REMOTE_VIEWONLY_PASSWORD" = "unset"; then
+    echo "generate viewonly password"
     REMOTE_VIEWONLY_PASSWORD=$(openssl rand 6 | python -c "import sys, base64; sys.stdout.write(base64.b32encode(sys.stdin.read()).lower())" | tr -d =)
     export REMOTE_VIEWONLY_PASSWORD
   fi
 
   if test -z "$REMOTE_READWRITE_PASSWORD" -o "$REMOTE_READWRITE_PASSWORD" = "unset"; then
+    echo "generate read/write password"
     REMOTE_READWRITE_PASSWORD=$(openssl rand 6 | python -c "import sys, base64; sys.stdout.write(base64.b32encode(sys.stdin.read()).lower())" | tr -d =)
     export REMOTE_READWRITE_PASSWORD
   fi
 
-  sed -i.bak "s/(password = WebUtil.getConfigVar\('password', ').+('\);)/\1$REMOTE_VIEWONLY_PASSWORD\2/g" /home/user/noVNC/index.html
-  if test -f /home/user/noVNC/index.html.bak; then
-    rm /home/user/noVNC/index.html.bak
+  if test -z "$REMOTE_AUTOMATIC_VIEW" -o "$REMOTE_AUTOMATIC_VIEW" = "true"; then
+    echo "integrate view password in index.html for noVNC"
+    sed -i.bak "s/(password = WebUtil.getConfigVar\('password', ').+('\);)/\1$REMOTE_VIEWONLY_PASSWORD\2/g" /home/user/noVNC/index.html
+    if test -f /home/user/noVNC/index.html.bak; then
+      rm /home/user/noVNC/index.html.bak
+    fi
   fi
 
   exec /usr/bin/supervisord -c /etc/supervisor/conf.d/openssh.conf "$@"
